+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Release History
 ---------------
 ## [Unreleased]
+### Fixed
+- Render or value function won't be called until table has data, see issue #32
+
 # [1.0.0-rc.10] - 2016-12-13
 ### Added
 - Ability to toggle column visibility, see issue #33